% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxidermist.R
\name{get_distinct_names}
\alias{get_distinct_names}
\title{Get distinct names from all rank columns.}
\usage{
get_distinct_names(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A data frame.
}
\description{
Get distinct names from all rank columns.
}
