% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxidermist.R
\name{gn_parse_names}
\alias{gn_parse_names}
\title{Parse names using gnparser.}
\usage{
gn_parse_names(input)
}
\arguments{
\item{input}{A vector of names.}
}
\value{
A data frame.
}
\description{
Parse names using gnparser.
}
