% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxidermist.R
\name{match_exact_worms}
\alias{match_exact_worms}
\title{Match scientificName with WoRMS and populate scientificName and scientificNameID.
Exact matches only, in case of multiple exact matches, the first one is used.}
\usage{
match_exact_worms(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A data frame.
}
\description{
Match scientificName with WoRMS and populate scientificName and scientificNameID.
Exact matches only, in case of multiple exact matches, the first one is used.
}
