% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxidermist.R
\name{parse_taxonomy}
\alias{parse_taxonomy}
\title{Add taxonomy to a data frame by parsing a taxonomy string column}
\usage{
parse_taxonomy(df, col)
}
\arguments{
\item{df}{A data frame.}

\item{col}{The column containing taxonomy strings formatted as 'tax=' followed by a comma-separated list of key:value pairs (e.g., d:Eukaryota,p:NA,...).}
}
\value{
A data frame.
}
\description{
Add taxonomy to a data frame by parsing a taxonomy string column
}
