% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxidermist.R
\name{parse_taxonomy_string}
\alias{parse_taxonomy_string}
\title{Parse taxonomy strings into a data frame}
\usage{
parse_taxonomy_string(input)
}
\arguments{
\item{input}{Character vector. Each element must contain 'tax=' followed by a comma-separated list of key:value pairs (e.g., d:Eukaryota,p:NA,...).}
}
\value{
A data frame with columns for each taxonomic rank and one row per input string.
}
\description{
Parse taxonomy strings into a data frame
}
