% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxidermist.R
\name{populate_species}
\alias{populate_species}
\title{Populate the species column using scientificName and taxonRank.}
\usage{
populate_species(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A data frame.
}
\description{
Populate the species column using scientificName and taxonRank.
}
