% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxidermist.R
\name{populate_worms}
\alias{populate_worms}
\title{Populate taxonomy fields using a numeric AphiaID column.}
\usage{
populate_worms(df, col = "AphiaID", replace_taxonomy = TRUE)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A data frame.
}
\description{
Populate taxonomy fields using a numeric AphiaID column.
}
