% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxidermist.R
\name{remove_unparsable_names}
\alias{remove_unparsable_names}
\title{Parse names in taxonomy columns using gnparser, remove unparsable names, and update scientificName and taxonRank.}
\usage{
remove_unparsable_names(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A data frame.
}
\description{
Parse names in taxonomy columns using gnparser, remove unparsable names, and update scientificName and taxonRank.
}
