% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxidermist.R
\docType{package}
\name{taxidermist}
\alias{taxidermist-package}
\alias{taxidermist}
\title{taxidermist}
\description{
Taxonomy tools.
}
\author{
\strong{Maintainer}: Pieter Provoost \email{pieterprovoost@gmail.com}

}
