% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxidermist.R
\name{update_scientificname_rank}
\alias{update_scientificname_rank}
\title{Update scientificName and taxonRank based on rank columns}
\usage{
update_scientificname_rank(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A data frame.
}
\description{
Update scientificName and taxonRank based on rank columns
}
